Move FilterBlock component to own file
